backend/requirements.txt
Flask==2.3.4
flask-cors==3.0.10
python-dotenv==1.0.0
pymongo==4.3.3
bcrypt==4.0.1
PyJWT==2.8.0
joblib==1.3.2
scikit-learn==1.3.2
